callbacks_pose_graph
